Remove footer skills
Already have list of skills per role. If that becomes too much at one
point, can move it back to the footer. Prefer having it by the role
because it remains accurate no matter how long ago the role was, since
it serves as a historical timeline. Compared to when skills are listed
in their own it's implied as current skills and has to be kept updated.