Add latest versions of resume and CV
These files are build artifacts so we shouldn't be staging them under
version control. It was fine in the early days and the size of the PDFs
are very small. But ultimately will need to build a CI pipeline that
builds the artifacts and uploads them to GitHub as assets.